Fixed errors with unsigned arithmetic

+++ specs.h
@@ -3,13 +3,13 @@
 #ifndef __SPECS_H
 #define __SPECS_H
 
-#include <stdint.h>
+// #include <stdint.h>
 
 #define GRANULARITY 10000000
 
 #define LATENCY (20 * GRANULARITY)
 
 // use 64-bit unsigned integers for time in nanoseconds, to avoid overflows
-typedef uint64_t timeunit;
+typedef unsigned long long timeunit;
 
 #endif
